F #9: refactor the verification role into smaller roles
This refactor was triggered by an image locked state, which
happened when we create the image, and run the core verification
services (that restarts services, and changes host states, etc.).
When we create, use and delete the image and only after that we
perform the core verifications, the image locked state does not
happen. to be investigated further...

Signed-off-by: Balazs Nemeth <bnemeth@opennebula.io>